+++ core_gearworks_lib/src/lib/engine/elems/.default/e_default.h
@@ -1,37 +1,14 @@
-#ifndef P3D_DEFAULT_H
-#define P3D_DEFAULT_H
+#ifndef E_DEFAULT_H
+#define E_DEFAULT_H
 
-// Include main header, but without the preset classes as they inherit from this class
-#define GW_NOINCLUDE_PRESET_DEFS
+#define GW_NOINCLUDE_ELEM_DEFS
 #include <gearworks.h>
-#undef GW_NOINCLUDE_PRESET_DEFS
+#undef GW_NOINCLUDE_ELEM_DEFS
 
-class defshape3d {
+class elem_default {
 private:
-	// This function applies any outstanding transformations to the shape.
-	// It is only ever run in the base render() function, and so it can be private.
-	// In order to publicly transform the primitive, use the translate(), scale(), and rotate() functions.
-	void apply_transformations();
-protected:
-	// Pointer to renderer object
-	gearworks::renderer *renderer_handle;
-
-	// IBO and VAO
-	gearworks::vertex_array_object vao;
-	gearworks::index_buffer_object ibo;
-
-	// Shape type:- (for some functions in render())
-	// | 0 = cube
-	unsigned int shape_type;
-
-	// Texture used for the primitive
-	gearworks::texture_2d *texture_object;
-public:
-	/// <summary>
-	/// A struct that holds properties of the shape. Don't reference this in your own program; use the shape::properties object instead.
-	/// </summary>
+	// A struct that holds properties of the shape.
 	struct shape_properties_struct {
-	public:
 		// Constructor for a shape_properties_struct object.
 		shape_properties_struct();
 
@@ -52,25 +29,18 @@ public:
 		/// <summary>
 		/// Basic transformation value of the shape.
 		/// </summary>
-		vec3 position, scale, rotation;
+		vec3 position, rotation, scale;
 	};
-
+protected:
+	// This function applies any outstanding transformations to the shape.
+	// It is only ever run in the base render() function, and so it can be private.
+	// In order to publicly transform the primitive, use the translate(), scale(), and rotate() functions.
+	virtual void apply_transformations(gearworks::renderer *renderer);
+public:
 	/// <summary>
 	/// The properties of the shape, like colour, texture, etc.
 	/// </summary>
 	shape_properties_struct properties;
-
-	/// <summary>
-	/// Default constructor. Note: to actually create the rectangle with vertex coordinates etc, use the create() function.
-	/// </summary>
-	defshape3d();
-	// Destructor
-	~defshape3d();
-
-	/// <summary>
-	/// Renders the shape via glDrawElements.
-	/// </summary>
-	void render();
 
 #pragma region Transformation methods
 	/// <summary>
@@ -119,4 +89,4 @@ public:
 #pragma endregion
 };
 
-#endif // header guard
+#endif