+++ core_gearworks_lib/src/lib/rendering/renderer/imp_r_renderer.cpp
@@ -6,7 +6,8 @@ gearworks::renderer::renderer() :
 	project_matrix(NULL),
 	view_matrix(glm::translate(mat4(1.0f), vec3(0, 0, 0))),
 	model_matrix(glm::translate(mat4(1.0f), vec3(0, 0, 0))),
-	current_zoom(1.0f)
+	current_zoom(1.0f),
+	fov(45.0f)
 {}
 gearworks::renderer::~renderer() {
 	DELETE_HALLOC(cur_window);
@@ -22,7 +23,7 @@ void gearworks::renderer::update_renderer() {
 	// Update the projection matrix so it is also resized along with window/viewport resize
 	// Also factor in zoom for zooming capabilities.
 	project_matrix = glm::perspectiveLH(
-		45.0f,
+		fov,
 		cur_window->get_winasprat(),
 		0.1f,
 		5000.0f